requirements.txt
av==10.0.0
imageio==2.31.1
imageio-ffmpeg==0.4.8
matplotlib==3.7.2
mmengine==0.8.2
natsort==8.4.0
opencv-python==4.8.0.74
PySimpleGUI>=4.60.5
PyYAML==6.0.1
torch>=2.0.0
torchvision
